% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_energy.R
\name{earthquakes_energy}
\alias{earthquakes_energy}
\title{Earthquakes energy calculation}
\usage{
earthquakes_energy(x, y = "erg")
}
\arguments{
\item{x}{A magnitude of earthquake in Richter magnitude scale}

\item{y}{A character to select units of energy: "erg" OR "joule"}
}
\value{
A number of the energy in the units
}
\description{
This function calculate the energy of a magnitude earthquake
}
\examples{
a<-earthquakes_energy(6, "erg")
}
