% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploteq.R
\name{ploteq}
\alias{ploteq}
\title{ploteq}
\usage{
ploteq(state1, year1)
}
\arguments{
\item{state1}{States of Mexico with its standard short name style}

\item{year1}{The year of earthquake until 2019}
}
\value{
a map interactive with notes in every point
}
\description{
Ploteq an interactive map with leaflet for earthquakes in every state
}
\examples{
ploteq("Veracruz", 2003)
}
